requirements.txt
paho-mqtt
requests
pyyaml
pydantic
ruamel.yaml
